Added test on deposit with WETH

+++ test/mocks/Tokens.sol
@@ -14,7 +14,7 @@ contract WETH is ERC20 {
         _mint(msg.sender, msg.value);
     }
 
-    function withdraw(uint wad) public {
+    function withdraw(uint256 wad) public {
         require(balanceOf(msg.sender) >= wad);
         _burn(msg.sender, wad);
         msg.sender.call{value: wad}("");
@@ -31,8 +31,8 @@ contract USDC is ERC20 {
     function mint(address account, uint256 amount) external {
         _mint(account, amount);
     }
+
     function decimals() public view override returns (uint8) {
         return 6;
     }
 }
-
